Add files via upload
YOLOv8 is still under heavy development. Breaking changes are being introduced almost weekly. We strive to make our YOLOv8 notebooks work with the latest version of the library. Last tests took place on 06.10.2024 with version YOLOv8.3.5.YOLOv8 can be installed in two ways - from the source and via pip. This is because it is the first iteration of YOLO to have an official package. It can be trained on large datasets and is capable of running on a variety of hardware platforms, from CPUs to GPUs.